more updates to suspend data and adl data
There's no need to define an object to hold data... it's free form data and up to the developer.

Suspend data doesn't need anything stored in the attempt's state object. We can just use the convention that suspend data will be keyed with the attempt's activity id, agent, state id of http://adlnet.gov/xapi/profile/scorm/types/adl-suspend-data, and optional registration uuid. 

ADL data is a lot like suspend data but there could be many and could span activities and attempts, which means the state properties need stored. So there's an array of objects with that information.